refactor : projection 객체에서 view 요소 삭제, MessageFormatter가 title, body를 결정하도록 로직을 분리 - client 코드가 구체적인 관련 코드(Message 관련)를 모르도록 AlarmService를 추상화 - 구체적인 Service 구현하기 전
Related to : #17
