--- src/main/java/com/umc5th/muffler/alarm/service/AlarmService.java
@@ -1,8 +1,0 @@
-package com.umc5th.muffler.alarm.service;
-
-import java.util.List;
-
-public interface AlarmService {
-    void sendAlarm(Alarmable alarmable);
-    <T extends Alarmable> void sendAlarms(List<T> alarms);
-}
