--- src/main/java/com/umc5th/muffler/alarm/service/Alarmable.java
@@ -1,8 +1,0 @@
-package com.umc5th.muffler.alarm.service;
-
-
-public interface Alarmable {
-    String getTitle();
-    String getBody();
-    String getToken();
-}
