+++ src/main/java/com/umc5th/muffler/domain/dailyplan/dto/DailyPlanAlarm.java
@@ -1,10 +1,12 @@
 package com.umc5th.muffler.domain.dailyplan.dto;
 
 import com.querydsl.core.annotations.QueryProjection;
-import com.umc5th.muffler.alarm.service.Alarmable;
+import com.umc5th.muffler.message.dto.Alarmable;
 import lombok.AllArgsConstructor;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+@Getter
 @AllArgsConstructor
 @NoArgsConstructor
 public class DailyPlanAlarm implements Alarmable {
@@ -18,20 +20,6 @@ public class DailyPlanAlarm implements Alarmable {
         this.memberName = memberName;
         this.dailyBudget = dailyBudget;
         this.alarmToken = alarmToken;
-    }
-
-    @Override
-    public String getTitle() {
-        return TITLE;
-    }
-
-    @Override
-    public String getBody() {
-        return "오늘의 소비 목표는 "
-                + dailyBudget
-                + "원 이에요! 오늘도 "
-                + memberName
-                + "님의 알뜰한 하루를 응원해요.";
     }
 
     @Override
