+++ src/main/java/com/umc5th/muffler/message/dto/Alarmable.java
@@ -1,0 +1,5 @@
+package com.umc5th.muffler.message.dto;
+
+public interface Alarmable {
+    String getToken();
+}
