+++ src/main/java/com/umc5th/muffler/message/service/AlarmService.java
@@ -1,0 +1,8 @@
+package com.umc5th.muffler.message.service;
+
+import com.umc5th.muffler.domain.dailyplan.dto.DailyPlanAlarm;
+import java.util.List;
+
+public interface AlarmService {
+    List<String> sendDailyAlarms(List<DailyPlanAlarm> dailyPlanAlarms);
+}
